+++ src/main/java/com/exult/ProjectCisco/service/Xde/XdeService.java
@@ -8,8 +8,8 @@ import java.util.Set;
 
 public interface XdeService {
     Set<Xde> findXde(String x);
-    Optional<Xde> findById(Integer x);
-    boolean deleteXde(Integer id);
-    Xde updateXde(Integer id, String name, Maven maven);
+    Optional<Xde> findById(Long x);
+    boolean deleteXde(Long id);
+    Xde updateXde(Long id, String name, Maven maven);
     Xde insertXde(String name , Maven maven);
 }
